5.1 - Case Study - Iteration 4 - Look Command/SwinAdventure/SwinAdventure/SwinAdventure/Player.cs
using System;
using System.Collections;
using System.Collections.Generic;
using System.Linq;
using System.Xml.Linq;

namespace SwinAdventure
{
    public class Player : GameObject, IHaveInventory
    {
        private Inventory _inventory;

        public Player(string name, string desc) :
            base(new string[] { "me", "inventory" }, name, desc)
        {
            _inventory = new Inventory();
        }

        public GameObject Locate(string id)
        {
            if (AreYou(id))
            {
                return this;
            }
            return _inventory.Fetch(id);
        }

        public override string FullDesciption
        {
            get
            {
                return $"You are {Name}, you are carrying\n{_inventory.ItemList}";
            }
        }

        public Inventory Inventory
        {
            get => _inventory;
        }

    }
}

